requirements.txt
pandas
requests
pyyaml
sqlalchemy
pyodbc
pytest
azure-identity
azure-keyvault-secrets
